requirements.txt
streamlit
facexlib
numpy
opencv-python
torch
Pillow
spandrel
spandrel_extra_arches